add lambda as argument to delete multiple datasources
